devops-project/backend/pyproject.toml
[project]
name = "devops-backend"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.34.0",
    "psycopg2-binary==2.9.10",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "httpx==0.28.1",
    "ruff==0.9.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
